Session

+++ app/Controllers/PaymentType.php
@@ -2,6 +2,7 @@
 
 namespace App\Controllers;
 use App\Models\PaymentTypeModel;
+use App\Controllers\PermissionChecker;
 
 class PaymentType extends BaseController
 {	
@@ -9,61 +10,74 @@ class PaymentType extends BaseController
 		helper('url');
         helper('form');
 		$this->paymentType = new paymentTypeModel();
+		$this->permcheck = new PermissionChecker();
 		$this->session = \Config\Services::session();
     }
 
 	public function index()
 	{	
-		date_default_timezone_set('Asia/Manila');
-        $data['payment_types'] = $this->paymentType->findAll();
-		return view('payment/type/index', $data);
+		if($this->session->has('logged_in') && $this->permcheck->check($this->session->get('id'), 'PaymentType')){
+			date_default_timezone_set('Asia/Manila');
+			$data['payment_types'] = $this->paymentType->findAll();
+			return view('payment/type/index', $data);
+		}
+		return view('errors/html/error_404');
 	}
 
 	public function add(){
-		date_default_timezone_set('Asia/Manila');
+		if($this->session->has('logged_in') && $this->permcheck->check($this->session->get('id'), 'PaymentTypeAdd')){
+			date_default_timezone_set('Asia/Manila');
 
-		if($this->request->getMethod()=='post'){
-			$data =[
-				'name' => $this->request->getVar('name'),
-				'description' => $this->request->getVar('description'),
-			];
+			if($this->request->getMethod()=='post'){
+				$data =[
+					'name' => $this->request->getVar('name'),
+					'description' => $this->request->getVar('description'),
+				];
 
-			if($this->paymentType->save($data) === true){
-				$this->session->setTempdata('success', 'Added Successfully!', 3);
-				return redirect()->to(base_url().'/payment/type');
-			} else {
-				$this->session->setTempdata('error', 'Adding Failed!', 3);
+				if($this->paymentType->save($data) === true){
+					$this->session->setTempdata('success', 'Added Successfully!', 3);
+					return redirect()->to(base_url().'/payment/type');
+				} else {
+					$this->session->setTempdata('error', 'Adding Failed!', 3);
+				}
 			}
+			return view('payment/type/add');
 		}
-		return view('payment/type/add');
+		return view('errors/html/error_404');
 	}
 
 	public function edit($id=null){
-		date_default_timezone_set('Asia/Manila');
-		$data['payment_type'] = $this->paymentType->find($id);
-		
-		if($this->request->getMethod()=='post'){
-			$data =[
-				'name' => $this->request->getVar('name'),
-				'description' => $this->request->getVar('description'),
-			];
+		if($this->session->has('logged_in')  && $this->permcheck->check($this->session->get('id'), 'PaymentTypeEdit')){
+			date_default_timezone_set('Asia/Manila');
+			$data['payment_type'] = $this->paymentType->find($id);
+			
+			if($this->request->getMethod()=='post'){
+				$data =[
+					'name' => $this->request->getVar('name'),
+					'description' => $this->request->getVar('description'),
+				];
 
-			if($this->paymentType->update($id, $data) === true){
-				$this->session->setTempdata('success', 'Updated Successfully!', 3);
-				return redirect()->to(base_url().'/payment/type');
-			} else {
-				$this->session->setTempdata('error', 'Update Failed!', 3);
+				if($this->paymentType->update($id, $data) === true){
+					$this->session->setTempdata('success', 'Updated Successfully!', 3);
+					return redirect()->to(base_url().'/payment/type');
+				} else {
+					$this->session->setTempdata('error', 'Update Failed!', 3);
+				}
 			}
+			return view('payment/type/edit', $data);
 		}
-		return view('payment/type/edit', $data);
+		return view('errors/html/error_404');
 	}
 
 	public function delete($id=null){
-		if($this->paymentType->where('payment_type_id',$id)->delete()){
-			$this->session->setTempdata('success', 'Deleted Successfully!', 3);
-			return redirect()->to(base_url().'/payment/type');
-		}else{
-			$this->session->setTempdata('error', 'Delete Failed!', 3);
+		if($this->session->has('logged_in') && $this->permcheck->check($this->session->get('id'), 'PaymentTypeDelete')){
+			if($this->paymentType->where('payment_type_id',$id)->delete()){
+				$this->session->setTempdata('success', 'Deleted Successfully!', 3);
+				return redirect()->to(base_url().'/payment/type');
+			}else{
+				$this->session->setTempdata('error', 'Delete Failed!', 3);
+			}
 		}
+		return view('errors/html/error_404');
 	}
 }
